thermistor code appears to be working, need to test with 3 therms

+++ science_nucleo/Core/Inc/thermistor.h
@@ -10,12 +10,12 @@
 #ifndef THERMISTOR_H_
 #define THERMISTOR_H_
 
-#include "stm32f3xx_hal.h"
+#include "stm32f3xx.h"
 #include "stdint.h"
 #include "stdlib.h"
 #include "stdio.h"
 #include "math.h"
-
+#include "string.h"
 
 ///////////////////
 //
@@ -69,6 +69,6 @@ void deleteThermistors(Thermistors*);
 ///////////////////
 
 // Returns raw data from the ADC pin given in a 12 bit string - DATA NOT FORMATTED
-uint16_t readVoltage(ADC_HandleTypeDef*);
+uint32_t readVoltage(ADC_HandleTypeDef*);
 
 #endif
